// packages/splits-sdk/src/subgraphv2/types.ts
type Scalars = {
  ID: string
  String: string
  Boolean: boolean
  Int: number
  Float: number
  DateTime: string
}

export type GqlRecipient = {
  id: Scalars['ID']
  account: GqlAccount
  ownership: Scalars['String']
}

type GqlAccountSharedFields = {
  chainId: Scalars['String']
  distributions: GqlTokenBalance[]
  latestBlock: Scalars['Int']
  latestActivity: Scalars['String']
  parentEntityType?: Scalars['String']
  internalBalances: GqlTokenBalance[]
  warehouseBalances: GqlTokenBalance[]
  contractEarnings: GqlContractEarnings[]
  warehouseWithdrawConfig: GqlWarehouseWithdrawConfig
}

export type GqlSplit = GqlAccountSharedFields & {
  __typename: 'Split'
  id: Scalars['ID']
  recipients: GqlRecipient[]
  type: 'split' | 'splitV2Pull' | 'splitV2Push'
  distributorFee: Scalars['String']
  distributeDirection: 'push' | 'pull'
  distributionsPaused: Scalars['Boolean']
  controller?: { id: Scalars['String'] }
  newPotentialController?: { id: Scalars['String'] }
  liquidSplit?: GqlLiquidSplit
}

type GqlContractEarningsBalance = GqlTokenBalanceSharedData & {
  contract: GqlAccount
}

export type GqlContractEarnings = {
  contract: GqlAccount
  internalBalances: GqlContractEarningsBalance[]
  withdrawals: GqlContractEarningsBalance[]
}

export type GqlUser = GqlAccountSharedFields & {
  __typename: 'User'
  id: Scalars['ID']
  withdrawals: GqlTokenBalance[]
  contractEarnings: GqlContractEarnings[]
}

export type GqlVestingModule = GqlAccountSharedFields & {
  __typename: 'VestingModule'
  id: Scalars['ID']
  vestingPeriod: Scalars['String']
  beneficiary: GqlAccount
  streams?: GqlVestingStream[]
}

export type GqlVestingStream = {
  id: Scalars['ID']
  streamId: Scalars['String']
  token: GqlToken
  startTime: Scalars['String']
  totalAmount: Scalars['Int']
  claimedAmount: Scalars['Int']
  account: GqlVestingModule
}

export type GqlWaterfallModule = GqlAccountSharedFields & {
  __typename: 'WaterfallModule'
  id: Scalars['ID']
  token: GqlToken
  tranches: GqlWaterfallTranche[]
  distributions: GqlTokenBalance[]
  nonWaterfallRecipient: GqlAccount
}

export type GqlWaterfallTranche = {
  id: Scalars['ID']
  startAmount: Scalars['Int']
  size?: Scalars['Int']
  claimedAmount: Scalars['Int']
  account: GqlWaterfallModule
  recipient: GqlAccount
}

export type GqlLiquidSplit = GqlAccountSharedFields & {
  __typename: 'LiquidSplit'
  id: Scalars['ID']
  distributorFee: Scalars['String']
  holders: GqlHolder[]
  split: GqlSplit
  isFactoryGenerated: Scalars['Boolean']
}

export type GqlHolder = {
  id: Scalars['ID']
  ownership: Scalars['String']
}
// ...
export type GqlSwapper = GqlAccountSharedFields & {
  __typename: 'Swapper'
  id: Scalars['ID']
  owner: GqlAccount
  paused: Scalars['Boolean']
  beneficiary: GqlAccount
  tokenToBeneficiary: GqlToken
  oracle: GqlOracle
  defaultScaledOfferFactor: Scalars['String']
  scaledOfferFactorPairOverrides: GqlSwapperScaledOfferFactorOverride[]
  swapperSwapBalances: GqlSwapBalance[]
}
// ...
export type GqlPassThroughWallet = GqlAccountSharedFields & {
  __typename: 'PassThroughWallet'
  id: Scalars['ID']
  owner: GqlAccount
  paused: Scalars['Boolean']
  passThroughAccount: GqlAccount
  passThroughSwapBalances: GqlPassThroughWalletSwapBalance[]
}

type GqlToken = {
  id: Scalars['ID']
  withdrawals: GqlTokenBalance[]
  internalBalances: GqlTokenBalance[]
}

export type GqlTokenBalanceSharedData = {
  id: Scalars['ID']
  amount: Scalars['Int']
}

export type GqlTokenBalance = GqlTokenBalanceSharedData & {
  id: Scalars['ID']
  account: GqlAccount
}

export type GqlWarehouseWithdrawConfig = {
  paused: Scalars['Boolean']
  incentive: Scalars['Int']
}
// ...
export type GqlAccount =
  | GqlUser
  | GqlSplit
  | GqlVestingModule
  | GqlWaterfallModule
  | GqlLiquidSplit
  | GqlSwapper
  | GqlPassThroughWallet

export type GqlAccountType = GqlAccount['__typename']

export type IAddress = `0x${string}`
export type IBalance = Dictionary<bigint>
export type ISplitType = 'immutable' | 'mutable' | 'liquid'

export type IAccount = {
  address: IAddress
  chainId: SupportedChainId
}

type IAccountSharedFields = IAccount & {
  distributions: IBalance
  latestBlock: number
  latestActivity: number
  parentEntityType?: string
  controllingSplits?: IAddress[]
  pendingControlSplits?: IAddress[]
  ownedSwappers?: IAddress[]
  ownedPassThroughWallets?: IAddress[]
  upstreamSplits?: IAddress[]
  upstreamWaterfalls?: IAddress[]
  upstreamLiquidSplits?: IAddress[]
  upstreamSwappers?: IAddress[]
  upstreamPassThroughWallets?: IAddress[]
  upstreamVesting?: IAddress[]
  contractEarnings: { [address: string]: IBalance }
  warehouseWithdrawConfig?: {
    paused: boolean
    incentive: number
  }
}

export type ISplit = IAccountSharedFields & {
  type: 'split' | 'splitV2Push' | 'splitV2Pull'
  balances: IBalance
  distributed: IBalance
  recipients: IRecipient[]
  distributorFee: number
  distributionsPaused: boolean
  distributeDirection: 'push' | 'pull'
  controller: IAddress
  newPotentialController: IAddress
  hash: string
  internalBalances: IBalance
  warehouseBalances: IBalance
  liquidSplitId?: IAddress
}

export type IRecipient = {
  address: IAddress
  ownership: number
  ens?: string
}